requirements.txt
pandas
psycopg2-binary==2.9.3
scikit-learn
imbalanced-learn
redis
numpy
alibi-detect
flask
prometheus_flask_exporter